requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.8.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cytoolz==0.12.2
dill==0.3.7
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
exceptiongroup==1.1.2
fastapi==0.100.0
flake8==6.0.0
frozenlist==1.4.0
h11==0.14.0
hexbytes==0.3.1
idna==3.4
isort==5.12.0
jsonrpcserver==5.0.9
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
lru-dict==1.2.0
mccabe==0.7.0
multidict==6.0.4
OSlash==0.6.3
parsimonious==0.9.0
platformdirs==3.9.1
protobuf==4.23.4
pycodestyle==2.10.0
pycryptodome==3.18.0
pydantic==2.0.3
pydantic_core==2.3.0
pyflakes==3.0.1
pylint==2.17.4
python-dotenv==1.0.0
pyunormalize==15.0.0
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.9.2
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
web3==6.6.1
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
